test: Add test for copyImageSync fn
