+++ package.json
@@ -32,7 +32,7 @@
     "npm-run-all": "4.1.5",
     "prettier": "2.5.1",
     "pretty-quick": "3.1.3",
-    "rimraf": "3.0.2",
+    "rimraf": "5.0.1",
     "ts-node": "10.4.0",
     "typescript": "4.5.2",
     "typescript-eslint-language-service": "4.1.5"
