+++ packages/reg-keygen-git-hash-plugin/package.json
@@ -32,7 +32,7 @@
     "glob": "7.2.0",
     "jest": "27.5.1",
     "reg-suit-interface": "^0.12.1",
-    "rimraf": "3.0.2",
+    "rimraf": "5.0.1",
     "ts-jest": "27.1.4",
     "typescript": "4.5.2"
   }
