+++ packages/reg-publish-s3-plugin/package.json
@@ -32,7 +32,7 @@
     "@types/mkdirp": "1.0.2",
     "@types/uuid": "8.3.4",
     "reg-suit-interface": "^0.12.1",
-    "rimraf": "3.0.2",
+    "rimraf": "5.0.1",
     "typescript": "4.5.2"
   }
 }
